identification: Set possible answers to same msgid
This patchs origin is https://lab.nexedi.com/nexedi/neoppod/merge_requests/20.
With recent changes on https://lab.nexedi.com/kirr/neo/merge_requests/2
and https://lab.nexedi.com/levin.zimmermann/wendelin.core/commit/f546d836ef3a95b0f2dfecacd4e74646af0c17e3
we can already use WCFS with multiple master nodes in combination with this
patch. Until we proceed in https://lab.nexedi.com/nexedi/neoppod/merge_requests/20
we can already use a NEO/py with this patch in order to test WCFS on our
project clone.

---

Original patch content:

When a node tries to connect to another node it initially sends a
'RequestIdentification' packet. The other node can either reply with
'AcceptIdentification' or in case of a secondary master with
'NotPrimaryMaster'.

In the second case the message id differs from the initial requests
message id. This makes it difficult in a multi-threaded implementation
to proceed this answer: due to the different msg/connection - id the
multi-threaded implementation tries to proceed this incoming message in
a different thread, while the requesting thread waits forever for its
peers reply.

The most straightforward solution is to use the same connection - id for
both possible answers to the 'RequestIdentification' packet. This
doesn't break given NEO/py implementation and is only a small patch.
A workaround in a multi-threaded implementation on the other hand
seems to be much more complicated and time-consuming. Finally it also makes
sense semantically, because "Message IDs are used to identify response
packets" and in the given context the 'NotPrimaryMaster' *is* the de facto
response of a 'RequestIdentification'.