Add branch for WC2 compatibility mode
Between NEO/go and NEO/py there are various incompatibilities.
This branch aims to transparently communicate those incompatibilities.
We can also specify a SR in SlapOS which uses this branch in case
someone aims to use WC2.

Here we start from the old-proto (e.g. pre-messagepack) branch, because
in NEO/go msgpack support isn't fully implemented yet. This isn't
considered as an incompatibility but simply as a remaining TODO on
NEO/go side.