client: Allow to force TLS via neos:// scheme
Similarly to how it is done with e.g. http:// and https:// - if neos://
is given TLS usage is forced and ca/cert/key must be there either in the
URI itself, or in $NEO_CA, $NEO_CERT and $NEO_KEY environment variables
mimicking the way how e.g. for https:// TLS credentials are taken from
host environment, not from the uri.

The latter might be usability convenience, but is also useful for WCFS
which needs to be able to remove secrets from uri on zurl normalization.

Please see discussion at https://lab.nexedi.com/nexedi/neoppod/merge_requests/18#note_184439
for details.

/cc @levin.zimmermann
/reviewed-by @jm
/reviewed-on https://lab.nexedi.com/nexedi/neoppod/merge_requests/21