reflink: add processing timestamp to transaction metadata in the reflink DB
Usually in ZODB, we can know when a transaction was committed by
decoding its TID, and by default it's the only way.

However in a reflink DB, transactions are committed with explicit TIDs,
which come from the main DB, for the following purpose: if you truncate
the main DB, just truncate the reflink DB at the same TID.

With this commit, it will be possible to know when reflink has processed
a TID by looking at transaction metadata.