reflink monitoring
It extends monitoring in the main DB to check whether a reflink process
is running and how late it is. For more information about the reflink
DB itself and in particular its backups if there are any, enabling
monitoring in the reflink cluster remains the only way.

See new options in admin & reflink.